flip() no longer requires a bind
The flip function is now defined in such a way as to keep the its
appropriate context in scope, absolving us of the need to bind it for
the requestAnimationFrame callback.

drawflag is now an integer array index to the canvas buffer rather than
a boolean that needs to be interpreted.